converted doubles to floats

+++ core/src/main/java/com/strategy_bit/chaos_brawl/ashley/components/CombatComponent.java
@@ -49,7 +49,7 @@ public class CombatComponent implements Component,Pool.Poolable {
         }
     }
 
-    public double getAttackDamage() {
+    public float getAttackDamage() {
         return attackDamage;
     }
